src/com/valentsolutions/xml/statepattern/providedstates/GenericByteValueState.java
package com.valentsolutions.xml.statepattern.providedstates;

import java.lang.reflect.Method;

import com.valentsolutions.xml.statepattern.HandlerState;
import com.valentsolutions.xml.statepattern.StateBasedContentHandler;

/**
 * 
 */
public class GenericByteValueState extends GenericValueState
{

    /**
     * @param handler
     * @param parentState
     * 
     * @pre handler != null
     * @pre method != null
     * @pre parentState instanceof method.getDeclaringClass()
     * @pre method.getParameterTypes().length == 1
     * @pre method.getParameterTypes()[0].getName() == "java.lang.Byte"
     */
    public GenericByteValueState(
		StateBasedContentHandler handler,
        StateBasedErrorHandler errors,
		HandlerState parentState,
		Method method)
	{
		super(handler, errors, parentState, method);
	}

    /**
     * 
     * @see GenericValueState#convertValueFromString(java.lang.String)
     * 
     * @pre value != null
     * @post return != null
     * @post return instanceof Byte
     */
    protected Object convertValueFromString(String value)
        throws ConversionException
    {
        try
        {
            return Byte.valueOf(value);
        }
        catch (NumberFormatException e)
        {
			throw new ConversionException(value, e);
        }
	}
}